Create README.md
I was fortunate to be selected among 36 Young Educators (Pengajar Muda) Indonesian Teaching Movement (GIM).
 
GIM is a movement—an effort to encourage everyone to 
participate in solving educational problems in Indonesia. Its vision is for all segments of society to be involved in the effort to educate the nation as a  promise of independence—a nation filled with leaders from various fields with global competencies and grassroots understanding.